PR #31 review — verplaats fbs-common-mappertests naar fbs-common (R21/R22)
R22: FbsCommonMapperTest in berichtenmagazijn testte grotendeels mappers
uit fbs-common. Tests over ProblemExceptionMapper, DomainValidationExceptionMapper
en IllegalArgumentExceptionMapper verhuisd naar libraries/fbs-common/src/test,
als pure JUnit 5 (geen @QuarkusTest), gesplitst per mapper:

- ProblemExceptionMapperTest (4 tests)
- DomainValidationExceptionMapperTest (1 test)
- IllegalArgumentExceptionMapperTest (1 test)

fbs-common/pom.xml: resteasy-core als test-scope dep toegevoegd voor de
JAX-RS RuntimeDelegate die Response.status(...) nodig heeft buiten een
Quarkus-context.

R21: IllegalArgumentExceptionMapperTest controleert nu op de exacte
gemaskeerde detail-tekst ("Er is een interne fout opgetreden...") i.p.v.
alleen assertNotEquals op de originele boodschap — veiliger tegen
regressies waarbij een andere lekkende tekst toch door de mapper komt.

De 4 tests over DbConstraintViolationExceptionMapper (berichtenmagazijn-
specifiek) blijven in berichtenmagazijn onder de nieuwe naam
DbConstraintViolationExceptionMapperTest. @QuarkusTest is daar behouden
zodat quarkus-jacoco de mapper-class instrumenteert voor coverage.

Coverage-config: fbs-common uitgesloten van de berichtenmagazijn-
coverage-check; fbs-common heeft een eigen testsuite en hoeft niet
dubbel gemeten te worden.
