+++ services/berichtenmagazijn/src/test/resources/application.properties
@@ -9,5 +9,5 @@ logboekdataverwerking.service-name=berichtenmagazijn-test
 
 # Uitgesloten: gegenereerde code en value-class-wrappers — die vertekenen
 # het coveragepercentage zonder testkwaliteit te meten.
-quarkus.jacoco.excludes=nl/rijksoverheid/moz/berichtenmagazijn/api/**
+quarkus.jacoco.excludes=nl/rijksoverheid/moz/berichtenmagazijn/api/**,nl/rijksoverheid/moz/fbs/common/**
